mean_amplitude_spectra(cardboard)
here are the new averaged (out of 5 "good" trials) amplitude spectra for our new 7 speeds, recorded over cradboard ontop the the glassplates.
i also added an older figure of "glass" mean amplitude spectra as comparison.
while the magnitude of amplitudes is not comparable between the 2 figures, the reduction in noisyness is apparent, due to the reduction in timespan of accelerometer data, and due to using only one interval instead of two.

next step is to use these spectra to generate input for the tractor, and measure the resulting acceleration on the fingertip (on glass) with no box movement. the aim is to estimate an "amplitudeMultiplier" for each speed to tune the tractor in a way so the resulting acceleration on the finger (from tractor action alone) is of a similar magnitude as the previously recorded acceleration from box movement (on cardboard) without tractor activity. for this i will use the "tractor_accmeter" and "envelope_vibro_LMT" scripts (to be updated soon