updated files for pilot testing
here is an update for the first complete pilot test run i did today, with the relevant files used:
"play_vibro.m" replaces the original "envelope_vibro" for the playing of the tactor input, and takes "sound_input_final.mat" as input.
"sound_input_final.mat" is the result of our calibration procedure, and each of its 14 collumns (for 7 speeds and 2 distances) represents the sound input of one of our stimulus conditions. it is loaded in "SSHMain_pilot.m"
"pilot_felix_170915_1_responses.mat" is the result file of todays complete run with 336 trials (2 runs of our complete set of 168 trials).
it took about 1h10min to complete (with very short brakes). it ran without crashes :), only on several trials near the end the forceplate seemed to not properly react to input, but went back to normal after about ~15-20secs.

notable changes to the main parts of the experiment:
"SSHMain_pilot.m"
save the result files for all pilot test runs in one single folder, as *.txt and *.mat .
added 5 additional training trials at the start, selected randomly, (responses are not recorded).
response keys are restricted to "1 or 2" on the keyboard (or numpad).
removed all accelerometer notions that remained from the calibration procedures.
added 2 new figures: 
"resetFig" to be shown while the box is moved to its starting posistion after each trial, so resetting is not confused with regualr interval movement. resetting is alsways done with a fixed speed of 2550 (standard stimulus speed) instead of the least used speed.
"toomuch" is shown in red if a participant pushes on the forceplate with too much force (voltage > 800). this aborts the current trial, resets the box position, and restarts the same trial.

for "SSHMove_pilot.m"
changed and optimised the trial/interval procedure
added functioning "reset finger" clause to make sure that there are no possible remaining sheer forces, from a previous interval or box resetting movement, effecting the fingertip beofre each interval. this is done by halting the computation procedure until the mean forceplate buffer has returned to near zero (i.e. the participant lifts his finger from the glass a little).
added a trail restart clause if the forceplate is pushed too strong.
added functionality so the appropriate sound input collumn is chosen from "sound_input_final.mat" for each interval.
adjusted delays.
removed all accelerometer notions that remained from the calibration procedures.

i hope this will be the base on which the experiement will run.