pilot control run results
yesterday i did the control pilot run (no tractor action, just movement). 
here is a figure comparing the answer tendency from the first pilot (with tractor) and the control run (no tractor).
the preliminary results in the answers seem very close to what we got in the pilot run with tractor action.
it seems tractor action has no dominant effect on the answering tendency, at least not when we regard the dataset containing all distance conditions.