updated data for the mixed condition
i redid the mixed condition with 336 trials, because as i saw in the results did not balance which of the intervals in the mixed condition had vibration and which didn't. formerly only the standard stimulus has had vibration, now standard and comparison both have vibration equal times, which allows to the vibration specific bias.