new pilot results fig.
figure with the data of all my pilot2 sessions (840 trials overall).
this time differentiating the crucial mixed condition.
it really looks like our desired effect is somewhat visible: vibration for the comparison stimulus (but not the reference stimulus) shifts the curve to the left, compared to the other mixed condition. 
this would mean tractor vibration ontop of movement induced vibration tends to make a movement feel faster.