(first) final version of the experiment to be used with naive participants
the only significant change consists of the adding of another distance condition. (now 4 instead of formerly 3 distance cond.),
formerly the mixed distance always gave the reference stimulus the long distance.
a new distance condition was added, so the the mixed distance trials are now balanced between reference and the comparison stimulus.
this inceases the overall trial length from 504 to 674. duration of the experiment will be around 2 hours.
the mixed vibration condition is still set to 0vs1 in amplitude, for later testing this might be adjusted to 0.2vs1 or 0.5vs1.